Update script
Modified the rooster file to remove an error in the timing functions.
Added some global variables to the main file to track the cloud data
and the user device screen width. Added a mapping functionality to the
google maps. Added a moving bouncing functionality to the cloud image.
Added a dynamic screen width setter for the article element's height.